build in correct repository for arch automagically
The repository selection algorithm of osc build is a bit weird.

First, if user specifies a repository explicitly on the commandline a build is
attempted in it. This part is sane.

Then repository name specified in config as build_repository, 'standard',
'openSUSE_Factory' and 'openSUSE_Tumbleweed' are tried in order. If repository
of the name exists build is attempted in it regardless of architectures it
supports. Finally, if none of these exists build is attempted in the last
configured build repository (in alhabetical order) if any.

This is suboptimal in cases when you have three factory repositories configured
as build targets so your pacakges ca be bilt for any architecture and osc
always picks the x86 one regardless of the architecture you are on.

This changes the behavior to
 - try to use repository specified on commandline if any
 - else use repository configured as build_repository if exists
 - else prepend 'standard', 'openSUSE_Factory' and 'openSUSE_Tumbleweed' to
   list of repositories sorted in reverse-alphabetical order. Try each in turn
   and if the repository supports selected architecture use it.

With this change if the old selection gave usable repository same is used after
the change. If there is usable repository for current architecture it is used
which was not the case before.

Co-Author: Marcus Huewe <suse-tux@gmx.de>